% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcRegWeight.R
\name{calcRegWeight}
\alias{calcRegWeight}
\title{Calculate multifuse regression weight (for two time series)}
\usage{
calcRegWeight(x, y)
}
\arguments{
\item{x}{time series 1; object of class \link{ts}}

\item{y}{time series 2; object of class \link{ts}}
}
\value{
A time series data frame with (1) overlapping x and y time series, (2) interpolated x and y time series (xi, yi), (3) x and y magnitude (xmag, ymag) and (4) multifuse regression weight (w).
}
\description{
Calculates the multifuse regression weight (inverse normalised magnitude) for two time series. A detailed description is given in Reiche et al. 2015 (Section 3.1.2 "Weighted time series correlation" and Section 3.1.2.1 "Deriviation of the time series regression weight").
}
\examples{
## load example data
data(tsexample.rda)
xts <- ndvi_ts #Landsat NDVI example time series
yts <- hv_ts   #ALSO PALSAR HV example time series

## calculate regression weight
wc <- calcRegWeight(xts,yts)
wc     #show time series data frame
wc$w   #show inverse normalised magnitude (multifuse regression weight)

## plot time series
plot2ts(xts,yts,lab_ts1="x",lab_ts2="y",title="Original x and y time series")
plot2ts(wc$x,wc$y,lab_ts1="x",lab_ts2="y",title="Overlapping x and y time series")
plot2ts(wc$xi,wc$yi,lab_ts1="xi",lab_ts2="yi",title="Interpolated overlapping x and y time series")
}
\author{
Johannes Reiche
}
\references{
Reiche, J., Verbesselt, J., Hoekman, D. H. & Herold, M. (2015): Fusing Landsat and SAR time series to detect deforestation in the tropics. Remote Sensing of Environment. 156, 276-293. DOI: 10.1016/j.rse.2014.10.001. \url{http://www.sciencedirect.com/science/article/pii/S0034425714003885}
}

