% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fusets.R
\name{fusets}
\alias{fusets}
\title{Univariate time series fusion using regresssion based prediction}
\usage{
fusets(xts, yts, xi, yi, weight)
}
\arguments{
\item{xts}{original x time series; object of class "ts"}

\item{yts}{original y time series}

\item{weight}{vector of weights (one for each correlation pair)}

\item{max_ewf}{maximum ewf to be considered)}

\item{steps}{steps at which the ewf is tested}

\item{plot}{optional plot the weight optimization plot. default=FALSE (no plot)}
}
\description{
Univariate time series fusion using regresssion based prediction
}
\author{
Johannes Reiche
}

