% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/multifuse.R
\name{multifuse}
\alias{multifuse}
\title{Optimzie regression weight}
\usage{
multifuse(xts, yts, ewf = 1, optimize = FALSE, max_ewf = 2, steps = 1,
  order = 1, alpha = 1, plot = FALSE, wpoints = FALSE)
}
\arguments{
\item{max_ewf}{maximum ewf to be considered)}

\item{steps}{steps at which the ewf is tested}

\item{plot}{optional plot the weight optimization plot. default=FALSE (no plot)}

\item{xi}{interpolated time series}

\item{yi}{interpolated time series}

\item{weight}{vector of weights (one for each correlation pair)}
}
\description{
The optimized regression weight is determined by detecting the minimum p-value for ewf (exponential weight factor, weight^ewf) starting from ewf = 0. In cse a local maximum p-value occurs, the ewf that results in the minimum p-value before the local maxima is selected.
}
\author{
Johannes Reiche
}

