% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot2ts.R
\name{plot2ts}
\alias{plot2ts}
\title{Plot two time series with different Y scales}
\usage{
plot2ts(ts1, ts2, ylim_ts1 = NULL, ylim_ts2 = NULL, lab_ts1 = "",
  lab_ts2 = "", col_ts1 = "black", col_ts2 = "blue", points = TRUE,
  title = "")
}
\arguments{
\item{ts1}{time series 1 (ts1); object of class "ts"}

\item{ts2}{time series 2 (ts2); object of class "ts"}

\item{ylim_ts1}{the y limits (x1, x2) for ts1. The default value is NULL, indicating that the ylim are set to the min and max value of ts1}

\item{ylim_ts2}{the y limits (x1, x2) for ts2. The default value is NULL, indicating that the ylim are set to the min and max value of ts2}

\item{lab_ts1}{a title for ts1}

\item{lab_ts2}{a title for ts2}

\item{col_ts1}{colour for ts1. The default value is "black"}

\item{col_ts2}{colour for ts2. The default value is "blue"}

\item{points}{add points at time series observations. The default value is TRUE.}

\item{title}{plot title}
}
\description{
Plot two time series with different Y scales
}
\author{
Johannes Reiche
}

