% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regressionWeight.R
\name{regressionWeight}
\alias{regressionWeight}
\title{Calculate probability for deforestation}
\usage{
regressionWeight(x, y)
}
\arguments{
\item{bayts,minN=NULL,maxN=NULL,maxt=NULL,chi_ts1=0.5,chi_ts2=0.5,start=NULL,end=NULL)}{}

\item{bayts}{"bayts" time series data frame}

\item{minN}{Minimum number of future observations to confirm the change; Default=NULL -> no minN}

\item{maxN}{Maximum number of future observations to confirm the change; Default=NULL -> no maxN}

\item{maxt}{Maximum time (years) to confirm the change; Default=NULL -> no maxt}

\item{chi_ts1}{Threshold of PChange at which the change is confirmed in case of the last observation is from for time series 2; Default=0.5}

\item{chi_ts2}{Threshold of PChange at which the change is confirmed in case of the last observation is from for time series 2; Default=0.5}

\item{start}{Start date of monitoring period; Default=NULL -> start is start of bayts time series}

\item{end}{End date of monitoring period; Default=NULL -> end is end of bayts time series}
}
\value{
Updated "bayts" time series data frame with changes if detected
}
\description{
Calculate probability for deforestation
}
\author{
Johannes Reiche
}

