% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/multifuse.R
\name{multifuse}
\alias{multifuse}
\title{Multi-sensor Time series Fusion (MulTiFuse) of two univariate time series}
\usage{
multifuse(xts, yts, ewf = 1, optimize = TRUE, ewf_max = 2,
  ewf_steps = 0.1, order = 1, alpha = 1, plot = FALSE,
  wpoints = FALSE)
}
\arguments{
\item{xts}{Univariate time series 1; object of class \link{ts}}

\item{yts}{Univariate time series 2; object of class \link{ts}}

\item{ewf}{Fixed exponential weight factor (ewf) in case no regression weight optimization is done. Default value is ewf=1 (simple weight). ewf=0 (no weight) and ewf=2 (squared weight)}

\item{optimize}{Regression weight optimization to maximise the statistical significance of the correlation and to take exceptional cases into account}

\item{ewf_max}{Maximum ewf up to which the regression weight optimization is considered. Default value is 2.}

\item{ewf_steps}{ewf_steps steps at which the ewf is tested. The default value is 0.1}

\item{order}{Degree of polynomial with which the relationship is modelled. default=1 (linear regression)}

\item{alpha}{Optional p-value threshold up to which fusion is performed. In case pvalue is > alpha no fusion is performed. Default value is 1.}

\item{plot}{Optional weight optimization plot and regression plot.}
}
\value{
(1) fused time series, (2) ewf, (3) pvalue and (4) rsquared
}
\description{
Pixel-based Multi-sensor Time series Fusion (MulTiFuse) of two univariat time series. "multifuse" wrapps the core functions of the MulTiFuse approach. First, a weighted timeseries correlation is performed (\link{calcRegWeight}). To maximise the statistical significance of the correlation and to take exceptional cases into account, correlation weight optimization is done (\link{optimizeRegWeight}) before the relationship of the two time series is modelled through a weighted regression analysis \link{wRsquared}. The optimized regression model is utilized in a second step for regression-based prediction of time series observation to fuse the SAR and optical time series (\link{regfuse}). The output is a fused time series consisting of the original x time series observations and additional x observations at the observation times of the y time series. A detailed description of the MulTiFuse approach can be found in Reiche et. al. 2015 (see reference below).
}
\examples{
## load example data
load("tsexample.rda")
xts <- ndvi #Landsat NDVI example time series
yts <- hv   #ALSO PALSAR HV example time series
plot2ts(xts,yts,lab_ts1="Landsat NDVI",lab_ts2="ALOS PALSAR HV [dB]")

## apply multifuse
xfus <- multifuse(xts,yts,plot=TRUE)

xfus[[1]] #fused time series
xfus[[2]] #ewf
xfus[[3]] #pvalue
xfus[[4]] #rsquared

##plot fused time series
plot2ts(xfus[[1]],yts,lab_ts1="Fused Landsat NDVI",lab_ts2="ALOS PALSAR HV [dB]")
}
\author{
Johannes Reiche
}
\references{
Reiche, J., Verbesselt, J., Hoekman, D. H. & Herold, M. (2015): Fusing Landsat and SAR time series to detect deforestation in the tropics. Remote Sensing of Environment. 156, 276-293. DOI: 10.1016/j.rse.2014.10.001. \url{http://www.sciencedirect.com/science/article/pii/S0034425714003885}
}

