% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optimizeRegWeight.R
\name{optimizeRegWeight}
\alias{optimizeRegWeight}
\title{Optimzie regression weight}
\usage{
optimizeRegWeight(xi, yi, weight, max_ewf = 2, steps = 1, order = 1,
  plot = FALSE)
}
\arguments{
\item{xi}{interpolated time series}

\item{yi}{interpolated time series}

\item{weight}{vector of weights (one for each correlation pair)}

\item{max_ewf}{maximum ewf to be considered)}

\item{steps}{steps at which the ewf is tested}

\item{plot}{optional plot the weight optimization plot. default=FALSE (no plot)}
}
\description{
The optimized regression weight is determined by detecting the minimum p-value for ewf (exponential weight factor, weight^ewf) starting from ewf = 0. In cse a local maximum p-value occurs, the ewf that results in the minimum p-value before the local maxima is selected.
}
\author{
Johannes Reiche
}

