% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optimizeRegWeight.R
\name{optimizeRegWeight}
\alias{optimizeRegWeight}
\title{Optimize regression weight}
\usage{
optimizeRegWeight(xi, yi, w, ewf_max = 2, ewf_steps = 0.1, order = 1,
  plot = FALSE)
}
\arguments{
\item{xi}{interpolated x time series}

\item{yi}{interpolated y time series}

\item{w}{vector of multifuse regression weights (one for each correlation pair)}

\item{ewf_max}{maximum ewf to be considered. default=2.}

\item{ewf_steps}{steps at which the ewf is tested. The default=0.1.}

\item{plot}{optional plot the weight optimization plot. default=FALSE (no plot)}
}
\value{
A list with (1) optimized ewf (ewf.optimized), (2) p-value found for optimized ewf (pvalue.optimized), and further paramter
}
\description{
Optimize multifuse regression weight. The optimized regression weight is determined by detecting the minimum p-value for ewf (exponential weight factor, weight^ewf) starting from ewf = 0. In case a local maximum p-value occurs, the ewf that results in the minimum p-value before the local maxima is selected. A detailed description can be found in Reiche et al. 2015 (Section 3.1.2.2 "Regression weight optimization")
}
\examples{
## load example data
load("tsexample.rda")
xts <- ndvi_ts #Landsat NDVI example time series
yts <- hv_ts   #ALSO PALSAR HV example time series

## calculate regression weight
wc <- calcRegWeight(xts,yts)
wc     #show time series data frame
wc$w   #show inverse normalised magnitude (multifuse regression weight)

## optimize regression weight
opt <- optimizeRegWeight(wc$xi,wc$yi,wc$w,max_ewf=10,steps=0.1,order=1,plot=TRUE)
}
\author{
Johannes Reiche
}
\references{
Reiche, J., Verbesselt, J., Hoekman, D. H. & Herold, M. (2015): Fusing Landsat and SAR time series to detect deforestation in the tropics. Remote Sensing of Environment. 156, 276-293. DOI: 10.1016/j.rse.2014.10.001. \url{http://www.sciencedirect.com/science/article/pii/S0034425714003885}
}

