% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wRsquared.R
\name{wRsquared}
\alias{wRsquared}
\title{Calculate weighted R-squared}
\usage{
wRsquared(x, y, weight = NULL, order = 1, plot = FALSE, xlab = "",
  ylab = "", ylim = NULL, xlim = NULL)
}
\arguments{
\item{x}{time series or vector (independent variable)}

\item{y}{time series or vector (dependent variable)}

\item{weight}{an optional vector of weights to be used in the fitting process. Should be NULL or a numeric vector. If non-NULL, weighted least squares is used with weights weights (that is, minimizing sum(w*e^2)); otherwise ordinary least squares is used. See also ‘Details’,}

\item{order}{degree of polynomial with which the relationship is modelled. default=1 (linear regression)}

\item{plot}{optional plot the correlation plot. default=FALSE (no plot)}

\item{xlab}{name for x (plot paramter)}

\item{ylab}{name for y (plot paramter)}

\item{xlim}{x limits (x1, x2). default=NULL (plot paramter)}

\item{ylab}{y limits (x1, x2). default=NULL (plot paramter)}
}
\value{
r.squared and p.value. See lm {stats} for description (plot paramter)
}
\description{
Calculate weighted R-squared from two input time series,using weighted least squares
}
\author{
Johannes Reiche
}

